Removed all ObjectBox code from every variant except the Play Store version.
