+++ core/build.gradle.kts
@@ -6,16 +6,11 @@ buildscript {
     mavenCentral()
     maven { setUrl("https://jitpack.io") }
   }
-
-  dependencies {
-    classpath(Libs.objectbox_gradle_plugin)
-  }
 }
 plugins {
   `android-library`
 }
 plugins.apply(KiwixConfigurationPlugin::class)
-apply(plugin = "io.objectbox")
 
 android {
   defaultConfig {
@@ -62,7 +57,6 @@ dependencies {
   debugImplementation(Libs.leakcanary_android)
 
   implementation(Libs.android_arch_lifecycle_extensions)
-  implementation(Libs.objectbox_kotlin)
   implementation(Libs.webkit)
   testImplementation(Libs.kotlinx_coroutines_test)
   implementation(Libs.kotlinx_coroutines_android)
