+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/modules/ApplicationModule.kt
@@ -25,8 +25,6 @@ import android.os.storage.StorageManager
 import dagger.Module
 import dagger.Provides
 import org.kiwix.kiwixmobile.core.DarkModeConfig
-import org.kiwix.kiwixmobile.core.data.remote.ObjectBoxToLibkiwixMigrator
-import org.kiwix.kiwixmobile.core.data.remote.ObjectBoxToRoomMigrator
 import org.kiwix.kiwixmobile.core.downloader.DownloadMonitor
 import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadManagerMonitor
 import org.kiwix.kiwixmobile.core.reader.ZimFileReader
@@ -53,14 +51,6 @@ class ApplicationModule {
   @Provides
   @Singleton
   internal fun provideBookUtils(): BookUtils = BookUtils()
-
-  @Provides
-  @Singleton
-  fun provideObjectBoxToLibkiwixMigrator() = ObjectBoxToLibkiwixMigrator()
-
-  @Provides
-  @Singleton
-  fun provideObjectBoxToRoomMigrator() = ObjectBoxToRoomMigrator()
 
   @Provides
   @Singleton
