+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/modules/MigrationModule.kt
@@ -1,6 +1,6 @@
 /*
  * Kiwix Android
- * Copyright (c) 2019 Kiwix <android.kiwix.org>
+ * Copyright (c) 2025 Kiwix <android.kiwix.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -15,15 +15,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package org.kiwix.kiwixmobile.core.dao.entities
 
-import io.objectbox.annotation.Entity
-import io.objectbox.annotation.Id
+package org.kiwix.kiwixmobile.core.di.modules
 
-@Entity
-data class RecentSearchEntity(
-  @Id var id: Long = 0L,
-  val searchTerm: String,
-  val zimId: String,
-  val url: String?
-)
+import dagger.Module
+
+@Module
+abstract class MigrationModule
