+++ core/src/main/java/org/kiwix/kiwixmobile/core/page/bookmark/adapter/LibkiwixBookmarkItem.kt
@@ -18,7 +18,6 @@
 
 package org.kiwix.kiwixmobile.core.page.bookmark.adapter
 
-import org.kiwix.kiwixmobile.core.dao.entities.BookmarkEntity
 import org.kiwix.kiwixmobile.core.page.adapter.Page
 import org.kiwix.kiwixmobile.core.reader.ZimFileReader
 import org.kiwix.kiwixmobile.core.reader.ZimReaderSource
@@ -69,20 +68,6 @@ data class LibkiwixBookmarkItem(
     title = title,
     favicon = zimFileReader.favicon,
     libKiwixBook = libKiwixBook
-  )
-
-  constructor(
-    bookmarkEntity: BookmarkEntity,
-    libkiwixBook: Book?
-  ) : this(
-    zimId = bookmarkEntity.zimId,
-    zimFilePath = bookmarkEntity.zimReaderSource?.toDatabase(),
-    zimReaderSource = bookmarkEntity.zimReaderSource,
-    zimName = bookmarkEntity.zimName,
-    bookmarkUrl = bookmarkEntity.bookmarkUrl,
-    title = bookmarkEntity.bookmarkTitle,
-    favicon = bookmarkEntity.favicon,
-    libKiwixBook = libkiwixBook
   )
 
   constructor(
