+++ objectboxmigration/src/main/java/org/kiwix/kiwixmobile/objectboxmigration/ObjectBoxMigrationHandler.kt
@@ -1,0 +1,34 @@
+/*
+ * Kiwix Android
+ * Copyright (c) 2025 Kiwix <android.kiwix.org>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package org.kiwix.kiwixmobile.objectboxmigration
+
+import org.kiwix.kiwixmobile.core.data.ObjectBoxDataMigrationHandler
+import org.kiwix.kiwixmobile.objectboxmigration.data.ObjectBoxToLibkiwixMigrator
+import org.kiwix.kiwixmobile.objectboxmigration.data.ObjectBoxToRoomMigrator
+import javax.inject.Inject
+
+class ObjectBoxMigrationHandler @Inject constructor(
+  private val objectBoxToRoomMigrator: ObjectBoxToRoomMigrator,
+  private val objectBoxToLibkiwixMigrator: ObjectBoxToLibkiwixMigrator
+) : ObjectBoxDataMigrationHandler {
+  override suspend fun migrate() {
+    objectBoxToRoomMigrator.migrateObjectBoxDataToRoom()
+    objectBoxToLibkiwixMigrator.migrateObjectBoxDataToLibkiwix()
+  }
+}
