+++ settings.gradle.kts
@@ -2,6 +2,9 @@ include(
   ":core",
   ":app",
   ":custom",
-  ":install_time_asset"
+  ":install_time_asset",
+  ":objectboxmigration",
+  ":defaultMigration"
 )
 rootProject.name = "kiwix-android"
+include(":objectboxmigration2")
